Identify complexity of a formula

A formula or fmls object may have several levels of complexity that need to be assessed. This affects how the formula may need to be expanded based on specific or special terms, such as a mediation, interaction, etc. It as designed for usage with the built-in fmls class and its corresponding tm objects, however, can be used for generic formulas if needed as well.

Usage

complexity(x)

Arguments

x

An object that is either a single formula, fmls or vector of tm objects

Value

A single integery value from 0:5

Rules of Roles

Roles may be given or assumed based on the formula definition, such as LHS objects being outcomes and RHS objects being predictors, particularly in a model-focused paradigm. This makes certain assumptions about the prescribed roles that a term can be given:

There can only be a single outcome term on the LHS.

The RHS can not contain outcome terms.

There can only be a single exposure term on the RHS.

interaction terms must be expanded against an exposure, such that both terms are present independently and as an interaction e.g. Y ~ X + I + X:I

There may not be > 1 mediator on the RHS

mediator terms may not exist alone on the RHS.

mediator terms on the RHS exist along with the corresponding exposure, which may confound the relationship between mediator and outcome.

mediator terms may exist on the LHS, but the exposure term must be on the RHS without other confounder terms.

A single meta term, such as strata, may only exist on the RHS.

Complexity

The complexity of a formula is derived from how well it follows the rules of roles, how informative the directionality of the terms are, and the number of terms on the LHS and RHS

Order	Rules	Directionality	LHS	RHS

0th	+	-	≤ 1	≤ 1

1st	-	-	1	1

2nd	+	+	1	≥ 1

3rd	-	-	1	≥ 1

4th	-	+	> 1	≥ 1

See also